+++ apps/frontend/src/components/admin/RevealKeyDialog.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import type { RevealKeyComponentProps } from '@/components/admin/types';
+import type { RevealKeyComponentProps } from '@/types/ui';
 import { RevealKeyDialogEvm } from '@/components/admin/RevealKeyDialogEvm';
 
 export default function RevealKeyDialog(props: RevealKeyComponentProps) {
